# JpegDecoder: post-process baseline spectral data per MCU-row 