… src/ImageSharp/Formats/Jpeg/Components/Decoder/JpegFrame.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/JpegImagePostProcessor.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/SpectralConverter.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/SpectralConverter{TPixel}.cs
… src/ImageSharp/Formats/Jpeg/JpegDecoderCore.cs
… src/ImageSharp/Image{TPixel}.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/DecodeJpegParseStreamOnly.cs
… tests/ImageSharp.Tests.ProfilingSandbox/Program.cs
… tests/ImageSharp.Tests/Common/NumericsTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/JpegDecoderTests.Images.cs
… tests/ImageSharp.Tests/Formats/Jpg/JpegDecoderTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/JpegImagePostProcessorTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/ParseStreamTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/SpectralJpegTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/SpectralToPixelConversionTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/Utils/JpegFixture.cs
… tests/ImageSharp.Tests/Formats/Jpg/Utils/LibJpegTools.ComponentData.cs
… tests/ImageSharp.Tests/Formats/Jpg/Utils/LibJpegTools.SpectralData.cs
… tests/Images/External/ReferenceOutput/JpegDecoderTests/DecodeBaselineJpeg_jpeg420small.png