### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Fixes #1597.

This also fixes #1692 simply because tests were failing with new architecture.

### Performance
No change. Baseline images should obviously be faster on the global scale due to lower memory footprint but in local benchmarks there's no change. And it's easy to explain: actual decoding code wasn't changed, only memory allocation and temporal buffers management.

_Note: PR benchmark is a bit faster than master branch (it's even faster than that because of a virtual call per each image stride). But I recall this false positive due to a really small image size. Most likely it's due to lesser memory footprint of spectral blocks allocation. I assume that at 4k resolutions there should be a little to no difference between them as actual code behind parsing/decoding wasn't touched._

|                             Method |            TestImage |     Mean |     Error |    StdDev | Ratio |
|----------------------------------- |--------------------- |---------:|----------:|----------:|------:|
| JpegDecoderCore.ParseStream master | Jpg/b(...)e.jpg [21] | 3.680 ms | 0.0152 ms | 0.0143 ms |  1.00 |
| JpegDecoderCore.ParseStream PR     | Jpg/b(...)e.jpg [21] | 3.474 ms | 0.0196 ms | 0.0174 ms |  0.94 |

Unfortunately, I don't have access to jetbrains memory profiler so I had a lot simplier check - custom debug allocator which simply prints all allocations. Here are dumps from the same image with same subsampling but different save methods - baseline and progressive:

*All small `byte` allocations are removed for clarity*

Progressive
```
Allocation of Rgba32[311400]
Allocation of Single[8448]
Allocation of Single[8448]
Allocation of Single[8448]
Allocation of Vector4[519]
Allocation of Block8x8[5016]
Allocation of Block8x8[1254]
Allocation of Block8x8[1254]
```

Baseline
```
Allocation of Rgba32[311400]
Allocation of Single[8448]
Allocation of Single[8448]
Allocation of Single[8448]
Allocation of Vector4[519]
Allocation of Block8x8[132] <- yay!
Allocation of Block8x8[33]  <- yay!
Allocation of Block8x8[33]  <- yay!
```

P.S. If somebody has some free time and access to jetbrains memory profiler - this is super welcome for the full picture!

### Tests & Legacy code
Everything is up and running. Old pipeline is removed.

### TODO
- [x] Fix existing tests
- [x] Tests for `SpectralConverter<TPixel>`
- [x] Remove legacy post processing code