@JimBobSquarePants 2021-07-12
```
Allocation of Block8x8[132] <- yay!
Allocation of Block8x8[33]  <- yay!
Allocation of Block8x8[33]  <- yay!
```

@br3aker That is so nice to see! This should make a huge difference in our memory allocation benchmarks!