@br3aker 2021-07-12
**Update**
Existing spectral libjpeg comparison tests are now working ccrretly (actually found a bug in decoding code thanks to it!)