@br3aker 2021-07-13
@JimBobSquarePants found an error in jpeg decoder reference output files which was even marked as a bug:
```
// BUG: The following image has a high difference compared to the expected output: 1.0096%
TestImages.Jpeg.Baseline.Jpeg420Small,
```

I've re-saved actual jpeg to png using photoshop and got this difference from my file and existing png file:
![image](https://user-images.githubusercontent.com/20967409/125440325-70e0c7f3-fa0d-4677-ab78-de67856d4389.png)

With this fix image passes tolerance test:
```
*** Jpg/baseline/jpeg420small.jpg ***
Difference: 0,2863%
```

I'm not sure how I can update it, can you elaborate please?