@br3aker 2021-07-13
**Update**
- Added integer ceiled division (tests included)
- Fixed parse stream only benchmark and updated results for PR imlementation (see top comment for comparison)
- Fixed (last) invalidated test which saves encoded
- Added docs to the `SpectralConverter` class
- Restored Sandbox code