@antonfirsov 2021-07-24
I did some WPA profiling on this on my 10 Core `i9-10900X`, and the results are quite impressive, although for some reason we didn't get the 8x memory footprint drop I expected. (Miscalculated the memory footprint of blocks in comparison to the image data?)

## 1. `LoadResizeSaveParallelMemoryStress` with `MaxDegreeOfParallelism = 8` and `Filter = JpegKind.Baseline`

#### Before the PR

Peak around 6.4 GB, total processing time `15.03` seconds.
![image](https://user-images.githubusercontent.com/6835152/126871904-01a31842-385f-463e-bd43-9505723874da.png)

#### Current master

Peak around 3.2 GB, total processing time `13.73` seconds.

![image](https://user-images.githubusercontent.com/6835152/126871928-0d376120-230b-407d-aec3-8a4aefa6b9d0.png)

## 2. `LoadResizeSaveParallelMemoryStress` with `MaxDegreeOfParallelism = 20` and `Filter = JpegKind.Any`

#### Before the PR

Peak around 13 GB.
![image](https://user-images.githubusercontent.com/6835152/126872005-4b1e1cb2-1390-43fa-98c7-750d48753e0e.png)

#### Current master

Peak around 4.4 GB.
![image](https://user-images.githubusercontent.com/6835152/126872042-633ab292-1bc1-426c-87d9-afbbc14176f3.png)

Great job @br3aker, thanks for the contribution!
/cc @JimBobSquarePants